Fixes installing/running everything as root user.
Neither should node packages be installed as root, nor should provision.sh be run as root user. This also removes the requirement for the generic vagrant-cachier cache buckets for the root user, which didn't work anyway. This also means, for now, that utilities (drush and drupal console) are NOT initialised anymore for the root user, which is the way they're intended to work. Provision.sh reflects these changes and is now invoked by the vagrant user.